requirements.txt
cryptography
qrcode[pil]
PyYaml
